refactor: 패널 resizable 로직 composable 분리
